ui(web): redesign settings page layout and controls
